refactor: 인터랙션 로직 통합 및 prop 전달 개선
